Move error definitions somewhere more sane
