+++ src/errors/index.ts
@@ -1,0 +1,5 @@
+export * from "./InvalidPonyFmUrlError.js";
+export * from "./InvalidYouTubeUrlError.js";
+export * from "./NotFoundError.js";
+export * from "./UnavailableError.js";
+export * from "./VideoError.js";
