+++ src/helpers/getPonyFmTrackInfoFromId.ts
@@ -1,5 +1,5 @@
 import fetch from "node-fetch";
-import { VideoError } from "../actions/getVideoDetails.js";
+import { VideoError } from "../errors/index.js";
 import { isObject, isString } from "./guards.js";
 
 // Based on https://github.com/Poniverse/Pony.fm/blob/a1522f3cd73d849099e4a3d897656dc8c4795dd7/app/Http/Controllers/Api/V1/TracksController.php#L129
